enhance order status page logic
