+++ lib/ui/pages/detail_order_history_page/detail_order_history_controller.dart
@@ -6,15 +6,14 @@ class DetailOrderHistoryController extends GetxController {
   void goToPage(OrderHistory orderHistory) {
     this.orderHistory = orderHistory;
     update();
-
-    Get.to(DetailOrderHistoryPage());
+    Get.to(DetailOrderHistoryPage(key: Key('DetailOrderHistoryPage')));
   }
 
   DetailOrderHistoryPage returnPage(OrderHistory orderHistory) {
     this.orderHistory = orderHistory;
     update();
 
-    return DetailOrderHistoryPage();
+    return DetailOrderHistoryPage(key: Key('DetailOrderHistoryPage'));
   }
 
   void exit() {
